JDK-8238535: Shape Clip does not work with AA rendering

+++ src/java.desktop/macosx/native/libawt_lwawt/java2d/metal/MTLTransform.h
@@ -15,7 +15,9 @@
                     M02:(jdouble) m02 M12:(jdouble) m12;
 - (void)resetTransform;
 
-- (void)setVertexMatrix:(id<MTLRenderCommandEncoder>)encoder destWidth:(jint)dw destHeight:(jint)dh;
+- (void)setVertexMatrix:(id<MTLRenderCommandEncoder>)encoder
+              destWidth:(NSUInteger)dw
+             destHeight:(NSUInteger)dh;
 @end
 
 #endif // MTLTransform_h_Included
